Update README.md
This repository is a collection of user interface examples created using the new and promising Flutter-inspired tool called 'flet' in Python.
Feel free to use the code shared in this repository for your own projects. If you find these examples helpful or interesting, please consider giving the repository a star. Your support and contribution motivate me to continue with the community.